refactor(database): split cleanup scan and generic cache for line-count ratchet
Extract action-item cleanup scan helpers and Redis path-keyed cache ops into
dedicated modules so action_items.py and redis_db.py stay under the 1500-line
product-file threshold without PR-body exceptions.

Verified: pytest test_open_action_items_count.py test_action_item_cleanup_* (60 passed).
